runtimes: change $262.IsHTMLDDA from a function to an object
This is how test262 expects it, which eshost willfully ignored since forever.

Tests relying on this only work with hosts that install their own
IsHTMLDDA (jsc, for example), which should be fixed by this change - at
least for the browser, jsshell, and kiesel runtimes, which actually
implement this.

See: https://github.com/tc39/test262/pull/2706#issuecomment-661589512